griefprevention integration bug fixes #2
Fixed a bug where whenever an admin is ignoring claims and tries to leash a mob not leashable by the vanilla game in someone else’s claim, it doesn’t leash. | Fixed a bug where when an admin is ignoring claims and leashes a mob that is leashable in the vanilla game, in someone else’s claim, it prints out the ‘in-claim-deny-mob message’ message, even though they can leash it.